+++ packages/frameworks/react/src/slider/slider-value-text.tsx
@@ -6,24 +6,24 @@ import { type Assign } from '../types'
 import { useSliderContext } from './slider-context'
 import { type UseSliderReturn } from './use-slider'
 
-export interface SliderOutputProps
+export interface SliderValueTextProps
   extends Assign<
     HTMLArkProps<'output'>,
     { children?: ((api: UseSliderReturn) => ReactNode) | ReactNode }
   > {}
 
-export const SliderOutput = forwardRef<HTMLOutputElement, SliderOutputProps>((props, ref) => {
+export const SliderValueText = forwardRef<HTMLDivElement, SliderValueTextProps>((props, ref) => {
   const { children, ...rest } = props
   const api = useSliderContext()
-  const mergedProps = mergeProps(api.outputProps, rest)
+  const mergedProps = mergeProps(api.valueTextProps, rest)
 
   const view = runIfFn(children, api)
 
   return (
-    <ark.output {...mergedProps} ref={ref}>
+    <ark.div {...mergedProps} ref={ref}>
       {view}
-    </ark.output>
+    </ark.div>
   )
 })
 
-SliderOutput.displayName = 'SliderOutput'
+SliderValueText.displayName = 'SliderValueText'