+++ packages/frameworks/solid/src/editable/editable.stories.tsx
@@ -1,35 +1,27 @@
 import type { Meta } from 'storybook-solidjs'
 import { Editable } from './'
-import { useEditable } from './use-editable'
 
-const meta: Meta = {
+type EditableType = typeof Editable
+
+const meta: Meta<EditableType> = {
   title: 'Editable',
+  component: Editable,
 }
 
 export default meta
 
-export const Hook = () => {
-  const api = useEditable({ placeholder: 'Enter' })
-  return (
-    <div {...api().rootProps}>
-      <div {...api().areaProps}>
-        <input {...api().inputProps} />
-        <span {...api().previewProps} />
-      </div>
-      {api().isEditing ? (
-        <>
-          <button {...api().submitTriggerProps}>Save</button>
-          <button {...api().cancelTriggerProps}>Cancel</button>
-        </>
-      ) : (
-        <button {...api().editTriggerProps}>Edit</button>
-      )}
-    </div>
-  )
-}
-
 export const Basic = () => (
-  <Editable.Root placeholder="Enter">
+  <Editable.Root placeholder="Placeholder" activationMode="dblclick">
+    <Editable.Label>Label</Editable.Label>
+    <Editable.Area>
+      <Editable.Input />
+      <Editable.Preview />
+    </Editable.Area>
+  </Editable.Root>
+)
+
+export const CustomControls = () => (
+  <Editable.Root placeholder="enter a value" value="Chakra">
     {(api) => (
       <>
         <Editable.Label>Label</Editable.Label>