+++ packages/frameworks/solid/src/select/select-control.tsx
@@ -1,4 +1,5 @@
 import { mergeProps } from '@zag-js/solid'
+import { Index } from 'solid-js'
 import { ark, type HTMLArkProps } from '../factory'
 import { useSelectContext } from './select-context'
 
@@ -8,5 +9,14 @@ export const SelectControl = (props: SelectControlProps) => {
   const api = useSelectContext()
   const mergedProps = mergeProps(() => api().controlProps, props)
 
-  return <ark.div {...mergedProps} />
+  return (
+    <>
+      <ark.div {...mergedProps} />
+      <select {...api().hiddenSelectProps}>
+        <Index each={api().collection.toArray()}>
+          {(option) => <option value={option().value}>{option().label}</option>}
+        </Index>
+      </select>
+    </>
+  )
 }