+++ packages/frameworks/solid/src/tooltip/tooltip.tsx
@@ -1,12 +1,15 @@
 import { type JSX } from 'solid-js/jsx-runtime'
 import { createSplitProps } from '../create-split-props'
+import { runIfFn } from '../run-if-fn'
 import { TooltipProvider } from './tooltip-context'
-import { useTooltip, type UseTooltipProps } from './use-tooltip'
+import { useTooltip, type UseTooltipProps, type UseTooltipReturn } from './use-tooltip'
 
-export type TooltipProps = UseTooltipProps & { children: JSX.Element }
+export type TooltipProps = UseTooltipProps & {
+  children: JSX.Element | ((api: UseTooltipReturn) => JSX.Element)
+}
 
 export const Tooltip = (props: TooltipProps) => {
-  const [useTooltipProps, restProps] = createSplitProps<UseTooltipProps>()(props, [
+  const [tooltipProps, localProps] = createSplitProps<UseTooltipProps>()(props, [
     'aria-label',
     'closeDelay',
     'closeOnEsc',
@@ -23,7 +26,8 @@ export const Tooltip = (props: TooltipProps) => {
     'positioning',
   ])
 
-  const api = useTooltip(useTooltipProps)
+  const api = useTooltip(tooltipProps)
+  const getChildren = () => runIfFn(localProps.children, api)
 
-  return <TooltipProvider value={api} {...restProps} />
+  return <TooltipProvider value={api}>{getChildren()}</TooltipProvider>
 }