+++ packages/frameworks/react/CHANGELOG.md
@@ -10,6 +10,12 @@ description: All notable changes to this project will be documented in this file
 
 - Added `FileUpload` component
 
+## [1.0.0-beta.4] - 2023-10-31
+
+### Fixed
+
+- Resolved an accessibility issue with `Select`
+
 ## [1.0.0-beta.3] - 2023-10-26
 
 ### Added